Update configurations to target CommonJS.
This commit will allow importing bare-server-node as a CommonJS package. A scenario that has led me to believe this change is necessary is the use of this package in CRA's (create-react-app) hook for setting up proxies. CRA only provides a CommonJS hook, which requires a weird workaround in order to register the Bare Server as a route in Express (before the dev server claims it) and then await a promise in the handler. This kind of syntax was not envisioned when bare-server-node was configured to target ESM.
This will not affect how the package is imported or required.